Support for non-literal seed in cp/mr/sp rand instructions, for datagen
Incl (1) tests, (2) fix missing test cases in testsuite 'misc', and (3)
fix reading scalar integers.

+++ src/main/java/com/ibm/bi/dml/runtime/util/MapReduceTool.java
@@ -306,7 +306,7 @@ public class MapReduceTool
 		br.close();
 		if( line == null )
 			throw new IOException("Empty file on hdfs: "+filename);
-		return Integer.parseInt(line);
+		return Long.parseLong(line);
 	}
 	
 	public static boolean readBooleanFromHDFSFile(String filename)
